backend/requirements.txt
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.9.0
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.3.0
black==23.12.1
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
fastapi==0.109.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.26.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
joblib==1.4.2
kubernetes==32.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.21.1
openai==1.9.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
overrides==7.7.0
packaging==25.0
pathspec==0.12.1
pgvector==0.2.5
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
posthog==4.0.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pulsar-client==3.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.1
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==7.4.4
pytest-asyncio==0.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
